finance-api/requirements.txt
# Core
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0          # async PostgreSQL driver
alembic==1.14.0          # database migrations

# Auth
python-jose[cryptography]==3.3.0
bcrypt==4.2.1

# Validation & Config
pydantic==2.10.2
pydantic-settings==2.6.1
python-multipart==0.0.12  # form data support

# Logging
structlog==24.4.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.28.0            # async test client
